requirements.txt
streamlit
google-genai
typer
python-dotenv